move shuffle into window

+++ core/src/main/java/org/apache/rocketmq/streams/core/rstream/GroupedStream.java
@@ -18,17 +18,14 @@ package org.apache.rocketmq.streams.core.rstream;
 
 import org.apache.rocketmq.streams.core.running.Processor;
 import org.apache.rocketmq.streams.core.runtime.operators.WindowInfo;
-import org.apache.rocketmq.streams.core.topology.virtual.GraphNode;
 
 import java.util.function.Supplier;
 
 public interface GroupedStream<K, V> {
 
-    GroupedStream<K,Long> count();
+    GroupedStream<K,Integer> count();
 
     WindowStream<K,V> window(WindowInfo windowInfo);
-
-
 
     GroupedStream<K,V> addGraphNode(String name, Supplier<Processor<V>> supplier);
